fix(#2755 R4): extract binding::worktree_state — binding.rs back under anti-monolith ceiling
The R4 item1 crash-after-bind adoption helper (WorktreeBindingState +
worktree_binding_state, +53 LOC) pushed src/binding.rs from 2489 to 2542,
over the 2500-LOC anti-monolith ceiling in tests/src_file_size_invariant.rs
(caught by the full nextest sweep; the R4 targeted filter had only exercised
the tight 750-LOC handlers ceiling in file_size_invariant.rs).

Extract the cohesive block into a `binding` submodule
(src/binding/worktree_state.rs) and re-export it, keeping the binding-format
coupling (parse_binding_guarded, the `worktree` field) inside the binding
module and the stable `crate::binding::worktree_binding_state` path unchanged
for both callers (checkout_recovery.rs, checkout_txn.rs). binding.rs → 2494.
No behavior change.


Agend-Agent: archfix-opus-3
Agend-Task: t-20260712233218081731-15764-2
Agend-Branch: fix/repo-checkout-transactional-submodules
Agend-Issued-At: 2026-07-13T06:49:34.424787+00:00
